[flow] Redesign the protocol for semantic decorations
Summary:
Initially, when I first introduced semantic decoration to power the refinement highlighting feature, I am deciding a pull vs push model in terms of how the client can get the data from the server.

The pull model does seem to be simpler, since the server just need to respond to the request, similar to how inlay hint is designed. However, if we want to use the push model, then we have to worry about pushing the information down, which will require us to hook into the complicated workflow that makes live errors work. Nevertheless, I underestimated the effort required to make the pull model to work:

- Every time something updates, we need to make vscode request new semantic decorations (easiest part)
- For various vscode actions (e.g. switching editors), we will need to re-request the information, otherwise there will be no decorations
- During initialization, we might send the request at the time when the server is not ready, and we somehow need to resend the request again.

The last point, which is a bug that exists right now, is the final problem that pushes the pull-based solution over the edge. I have tried multiple solutions but all failed. If I tried harder I can potentially make it work, but there might be just another issue I'm aware of. On the other hand, the push based model isn't that hard to do.

This diff implements the push based model. On the client side, we put `semanticDecorations` into init options. If the server sees that a client has this option enabled, then it will send down semantic decorations as diagnostics. On the client side, we will use middleware to intercept these special diagnostics, filter them away so that they are not visible, and store them to a separate hiddenDiagnostics map. Finally, we use the same infra of ANSI decorating error messages in detailed error rendering, to apply the semantic decorations to js files.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D61946960

------------------------------------------------------------------------
(from a6e57cdc65a6e7b3f9d3b7e072de1f8f789f42f3)

fbshipit-source-id: d27da49d10c2490307e8c93a58fdbeab48f05c1b