[flow][nat-inf] new exported Object.freeze behavior (under flag)
Summary:
Applies the same approach on Object.freeze in statement.ml to type_sig. In short:
```
Object.freeze({f:"a"});      // {+f: "a"}
Object.freeze({f:{g:"a"}});  // {+f:{g:string}}
Object.freeze({...{f:"a"}}); // {+f: "a"}
```

The changes are limited to type_sig_parse to keep the change syntax-based as it is in statement.ml.

Given that large error diff expected, this change is behind the `natural_inference.object_freeze` flag.

It also uses a fresh parameter `new_frozen` passed down various functions in type_sig_merge, to avoid interference with the old approach.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D59565113

fbshipit-source-id: e4736b4a7fcc313bed71e88505c7d47a9d124f07