[flow][website] Prepare for richer type-at-pos results
Summary:
In this next diff, I will add support for richer type-at-pos results that include refinement information. It will require the js side to handle an array of results, which might include either flow or markdown.

This diff makes the change on the js side first, so it can be easily tested that it still works for the old flow.js builds.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62137721

fbshipit-source-id: aabf5514f33fb1d43287641002fdc39fd5a1c974