[flow][tests] Get rid of $ObjMap in some tests 2/n
Summary:
We are going to kill $ObjMap, so let's starting to get rid of it in our tests first.

This diff updates the relay test. The type in this test has long diverged from the actual type in relay. This diff does some best effort rewrite to mostly keep the same errors.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62219061

fbshipit-source-id: 67c5fd1bc0da37d9493b58d74c26e6df6ccf3076