[flow] Retain all special hidden errors in LspErrors.limit_errors
Summary:
The semantic decoration feature, which currently publishes information to the client through the diagnostics channel, has some bad interactions with the server side error limit. In a giant file, the number of refined reads can exceed 200, which can cause no errors to be shown for the file.

For now, I decide to quickly unbreak it by always keeping these special diagnostics. Longer term, maybe we should re-design the protocol again to not make it go through diagnostics...

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62222992

fbshipit-source-id: 37483208d4d95907aa83f114cd5a10990b1d4c52