[flow][tests] Get rid of $ObjMap in some tests 3/n
Summary:
We are going to kill $ObjMap, so let's starting to get rid of it in our tests first.

In this diff, I removed the tests that already have a mapped type version, and replaced $ObjMap with mapped type if the tests are not already testing mapped type.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62220761

fbshipit-source-id: 3fc7c063d731a2c87fa65daeb27d090ea0fdb167