[flow][tests] Get rid of $ObjMap in some tests 4/n
Summary:
We are going to kill $ObjMap, so let's starting to get rid of it in our tests first.

In this diff, I removed the tests that already have a mapped type version, and replaced $ObjMap with mapped type if the tests are not already testing mapped type.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62245044

fbshipit-source-id: c0f0aee3cc6570261563cf4119ac3023981e2ab7