[flow] Make the duplicate component props be in alphabetical order
Summary:
Right now it's in reverse alphabetical order because we iterate over name map and ocaml do it in normal order and we build up a reversed list.

Let's reverse the list so that it's in usual order in the error, and also in the Omit keys we provide in quickfix.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D62242629

fbshipit-source-id: 6f344310bb33d3a437c3b4d9bdbbcee11bf6aac9