[flow] store number_literal in UnionEnum.Num
Summary:
One of the variants of an enum union is a number, which was represented as `Num of float`. The `float` is an internalization of the raw value of the number in the source. So, for example `1`, `1.0`, `1.00`, etc. all correspond to the same number.

The problem with this is that when we want to recover a type based on this union member we always recover the same serialization which is `1.000000`, instead of, for example, the more compact `1`. This is not a huge problem today since we don't often insert these values in errors, but will be also used in a follow-up diff.

This diff changes this by including the "raw" form of the number in the enum member. This way we can recover the same exact form we got from the source (see change in error diff in unit tests).

We need to be careful to ensure comparison and equality remain the same with earlier.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D62225799

fbshipit-source-id: ff49885777936ddf66a7a7336b1337dbdc55a15c