[flow] better errors for union-to-union enum mismatch
Summary:
On union-to-union mismatches, e.g. in
```
type T1 = 'a' | 'b' | 'c' | 'a1' | 'b1' | 'c1' | 'd1' | 'e1' | 'f1' | 'g1' | 'h1';
type T2 = 'a' | 'b' | 'c' | 'a2' | 'b2' | 'c2';
_ as T1 as T2; // error
```
the error message will now give a short list of the members that are not missing from the LHS and the number of the remaining missing members. E.g.:
```
Cannot cast `_` to `T2` because `T1` [1] is incompatible with `T2` [2]. Type `T1` [1] includes members `a1`, `b1`, `c1`
and 5 more that are not included in type `T2` [2]. [incompatible-cast]
```
Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D62270164

fbshipit-source-id: 56c174427fd30ab5f02fb4d76ffa73cfd7f580be