[flow] Add more information to UnionEnum
Summary:
Adds a tag to describe the kind of values included in an enum union. This information is useful given that `canon` will match various types to the same UnionEnum.t. In some cases (see following diffs that use negative result of `union_optimization_guard`), we will want to know if the initial DefT types were actually of the same kind.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D59768552

fbshipit-source-id: 0978a9f25b87267cf49ae8b4bed9772d401b5243