[flow][refactor] Extract current non RenderT ~> RenderT handling code into renders_kit
Summary:
I want to use this code in some form to power the code action to insert render type declaration automatically. Having it in subtyping_kit complicates things and it's hard to reason about how it interact with other rules. Fortunately, the entire continuous block of the subtyping logic can be moved here in a straightforward way.

It also makes me happy that most of the core render type subtyping logic is in one place now.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62273377

fbshipit-source-id: bfac305f7dd0806aeeca63a8032a7d98c022a923