[flow] raise a single error on union-to-singleton type mismatch
Summary:
`UnionT(enum) ~> SingletonStrT` constraints are only valid when all members of the enum are exactly the same as `SingletonStrT`.

This diff leverages this fact to produce fewer errors.

To avoid complications with implicit coercions that could make this constraint true, we make sure that the tag of the enum in the LHS is the same as the tag of the literal type in the RHS (similar to D59709663).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D62284942

fbshipit-source-id: 162b57cd6413af221b98a3d9620ee5db154a08af