[flow][refactor] Refactor implementation of normalizing non-generic render type to prepare for inserting render type code action
Summary:
Later, I want to use some of the same code to power a code action to insert render type declaration, since it already has some logic to infer the user intention and use the right render variant.

This diff mostly reorganizes the code into a module, and ensure that we emit all errors at the end. The latter is important, since for the code action use case, we don't want to actually emit the error but only to use it as a guide to decide what to put as render type.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62278126

fbshipit-source-id: 437a9f2d974e17088df7345237bcda76946a2f78