[flow] Support arrays and tuples for homomorphic mapped type
Summary:
This is surprisingly easy to do: we only need to intercept the case of `ArrT ~> ObjMap (homomorphic mapped type)`.

For now, I did the bare minimum support so that $TupleMap can be completely replaced. Notably, it doesn't allow you to change variance for now.

Changelog: [feature] Mapped type now supports array and tuple types in the form of `{[K in keyof <array or tuple type>: <mapped type>}`.

Reviewed By: panagosg7

Differential Revision: D62473783

fbshipit-source-id: 96db4043ea9614c0ff36b247de6a809626780c98