[flow] Accurately model the absence of component instance in component syntax components
Summary:
In this diff, I changed the internal representation of `ReactAbstractComponentT` to accurately model the absence of ref prop component syntax components.

For component syntax components that lack the ref prop, we will use `ComponentInstanceOmitted` to model it, and during subtyping, it will be interpreted as `void`, instead of `mixed`.

The goal is to enable us to reliably tell during hover whether we should display the ref prop, and also in the future enable us to accurately model optional ref prop.

Changelog: [errors] Component syntax components without ref prop will have `void` as the instance, so `React.ElementRef<component>` of such component will return void.

Reviewed By: jbrown215

Differential Revision: D62511547

fbshipit-source-id: 39e3ecf4c8605901067ca340a96dfc02bd41a6ba