[flow][libdef] Update core libdefs to use mapped type on array instead of $TupleMap
Summary:
This diff changes the core libdef to stop using `$TupleMap` and instead use the mapped type on array inputs.

`$TupleMap` currently has a hacky support for iterables by first translating them into array type. I decide not to do that, but instead use conditional type to model the return: first checking for array input and use the more precise mapped type if we can, and then fall back to the general iterable typing where the extra power of $TupleMap is not needed.

Changelog: [libdef] `Promise.all` and `Promise.allSettled` have been updated to use mapped type instead of `$TupleMap`. The requirement on the type arguments are slightly changed.

Reviewed By: panagosg7

Differential Revision: D62473784

fbshipit-source-id: b819359e7a9be3626b82c995e2f5a074ca7ac121