[flow] No longer wrap special TypeT around mapped type's prop type
Summary:
The implementation of mapped type is written before we have support for `from_value` in TypeApp, which will just do the substitution without unwrapping TypeT first. We should just use the new infra and avoid the special TypeT.

I need to change the mk_typeapp logic to not do repositioning to keep the current error location. It ends up improving the error for `typeof Poly<string>` as well. Note that we shouldn't do the reposition there in the first place, since it will reposition it back to the TypeApp location instead of the actual underlying type computed after `SpecializeT`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D62456874

fbshipit-source-id: 2bf6b36b411f94e17f18830793c0b22376ddd3df