[flow] Use React.RefSetter as the hint for the ref prop
Summary:
We should kill React.Ref since it's the confusingly named bad type that allows string ref that will be killed in React 19. Before we can do that, we need to get rid of all references of it in our ocaml code.

In this diff, we stop using it for hint decomposition. This is a straightforward change, since the addition of number and string is never going to be helpful for contextual typing. The same hint is also used to power hover on ref prop, so user will also stop seeing React$Ref on hover and stop copying the bad type.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D62539802

fbshipit-source-id: 8f22843c231c133acaf462cd2575bd99f66c5253