[flow] Support mapped type on generic arrays
Summary:
This diff intercepts the mapped type on generic array type case, so that it doesn't go to normal ObjectKit and ends with array is not object error.

In this case, we handle generics similar to how it is handled in slice_utils.

Changelog: [fix] Mapped type on generic arrays is now supported. Previously it will fail with array is not an object error.

Reviewed By: panagosg7

Differential Revision: D62770468

fbshipit-source-id: 7fc367cfd605c74a4860b2549c51cd91698086ed