[flow] better contextual typing for conditional expressions
Summary:
Before this change Flow would only use the type of the consequent branch of a conditional for hints to use in the alternate branch. So in
```
cond ? exp1 : exp2
```
it would use the type of `expr` as a hint fot `expr`.

It would not use the type of `expr2` to type `expr1` in cases like
```
cond ? [] : [1]
```
This diff changes his by using the `expr2`'s type when its type is synthesizable.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D62588523

fbshipit-source-id: 392397a998a13a398f7286adf6228d2bdadacdc1