[flow][ide] Code action to insert inferred render types
Summary:
This diff adds support for inserting inferred render types.
The normalization strategy mostly mirrors how render type subtyping work, and we will visit the type and gradually widen the inferred state to be less specific.

I was trying to reuse some of the code to autofix bad render types, but there is enough difference between two render type normalization mode (the type annotation one is more restrictive, while the promotion one is more permissive), so that there is not much to share.

Changelog: [ide] We now offer a code action to insert inferred render type when you hover on the name of the component

Reviewed By: panagosg7

Differential Revision: D63359493

fbshipit-source-id: 3566c2ee34c3f89fe0f02a6876998bb375b821a7