[flow][parser][component-syntax] Allow trailing comma after rest param
Summary:
There is no reason why we don't want to allow that, other than keeping consistent with the behavior of parsing function rest params. We can't change the JS spec for function rest params, but we can for component syntax to make it more permissive.

Prettier will still remove the trailing comma if we still care about the consistency of checked-in code, but it will at least make the experience of manually convert function component to component syntax slightly easier.

Changelog: [parser] Trailing comma is now allowed after rest parameter in component syntax.

Reviewed By: panagosg7

Differential Revision: D63488905

fbshipit-source-id: 8f946256ac8c529b5109cb19f5d5e755cff4da42