[flow][component-syntax] Switch to trust-but-verify model for the ref prop in component syntax component and types
Summary:
This diff fundamentally changes how the ref prop is handled for component syntax components and component type.

Currently, it follows the approach of using a builtin EvalT to extract the instance out of `React.RefSetter`, so that it can fit into the current `React.AbstractComponent` type. From the previous diff, we can already see that this approach makes implicit instantiation wrt the ref prop having weird results, and it will be a blocker for switching to react 19 model of ref-as-props.

This diff switches it to be trust-but-verify model, for the ref prop in component syntax and component type. We directly record the ref prop type as is using `ComponentInstanceAvailableAsRefSetterProp`, as opposed to `ComponentInstanceAvailableAsInstanceType`. We will do the same post inference check at the declaration site, but otherwise we will just use the original type and make no attempt to transform it, even if it's wrong. In various use sites, especially during subtyping, we will wrap the previous instance type with `React.RefSetter` if it needs to interact with the `ComponentInstanceAvailableAsRefSetterProp` variant. Under this scheme, implicit instantiation just works, similar to how implicit instantiation just works with all other ref props.

There is only one special case I did to have better errors in most cases: when `ComponentInstanceAvailableAsRefSetterProp` interact with other instance that's not `ComponentInstanceAvailableAsRefSetterProp`, we will still attempt to extract the instance type, if the ref prop has fully resolved type and we can extract. In this way, we will still run a smaller instance ~> instance flow that tends to have better errors that don't involve the union type in `React.RefSetter`.

Changelog: [errors]
* Component syntax component and component types' ref prop now must have `React.RefSetter<...>` type.
* String refs are no longer supported on component syntax components

Reviewed By: jbrown215

Differential Revision: D63148526

fbshipit-source-id: 8188730a17c969ed83aa740bd7c0ddf7240aac1f