[flow] natural_inference.object_freeze on by default
Summary: Changelog: [errors] The inferred type for `Object.freeze({ A: 'a', B: 'b' })` is now `{+A:'a',+B:'b'}` both locally within a file and when the object is being exported. This replaces an earlier unsound behavior where the type of A would opportunistically behaved either as `string` or `'a'` depending on the context where it appeared.

Reviewed By: SamChou19815

Differential Revision: D63679710

fbshipit-source-id: 15da407da393e3797857825ec75bc28d1c9dc00c