[flow] fix in unary arithmetic negation
Summary:
Typing the literal `-1` was inconsistent. In statement.ml it was inferred as the literal type `-1`, when not in a "frozen" context, by flipping the sign of the literal `1`. However, when it appeared under an `Object.freeze()` property we would not infer it as a `SingletonNumT (-1)`.

This diff fixes this behavior by propagating the `~frozen` flag to `unary`, and handling the `SingletonNumT` case in addition to `NumT`.

It also fixes the behavior in type_sig by applying the same transformation in `flow_unary_arith` which is called from annotation_inference.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D63679709

fbshipit-source-id: b9ec4ce97341301ceaa395a2b26e4c4e543f2cf2