[flow][ide] Support render type autocompletion for values with component type
Summary:
Currently, we only consistently do a good job of autocompleting the names in render types if the names point to component declarations. If it's not declarations, then we almost get nothing. The problem affects both local and remote values that need to be auto-imported. In this diff, I fixed it in both.

For local ones, it's actually fairly easy: in places where we look for `Ty.ComponentDecl`, we also match on `Ty.Type (Ty.Component)`. I added additional match to ensure that the component type has render declarations, so that it has a better chance to be actually a valid argument of render type operator.

For remote ones, I changed the exports indexing to include `ComponentAnnot` as well (surprise! React.AbstractComponent and the new component types are represented differently in type-sig). Note that it doesn't quite fix the issue when the abstract component is behind a type alias (as shown by the tests), but for the future when we completely killed React.AbstractComponent in favor of the component type, it should be good enough.

Note that there is a slight difference between local and remote render type completion in that the local one tries to ensure that you only use the component as a type for render type purposes. I intend to relax this after we make the shorthand syntax no longer be `React.Element` but another opaque type (like what D60634061 would do, but instead of interpreting them as component type, interpreting them as `declare opaque type React$RendersExactly<Component>: React.Node`).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D63728963

fbshipit-source-id: 08a96b6f653404f2e94d2401e5874b275a7a3d56