[flow] propagate frozen flag to Type_sig_merge.merge_obj_value_prop
Summary:
In
```
exports.numberValue1 = 42;
exports.numberValue2 = 42;
exports.numberValue3 = 42;
exports.numberValue4 = 42;
exports.stringValue = "str";
```
the object representing the exports needs to be non-writable.

This information is encoded in the "frozen" flag of the created object. This was not propagated to `merge_obj_value_prop`, and so the non-writable property was not enforced.

This diff fixes this.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D64131343

fbshipit-source-id: 791a23d34754ec69820c3db75b0dbf2fa7a4fafc