[flow][easy] update test comment
Summary:
When importing from a file A.js that includes
```
exports.foo = 42;
```
and accessing a non-existing property
```
var A = require("./A");
A.bar = 123;
```
before D63679710, we used to get an error that `A` is frozen, without attempting to look up the specific property. Now that "frozen" is not a property of the object type any more, we will just error for the fact that "bar" is not there.

This diff just updates the wording in the test comment to reflect that.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D64131342

fbshipit-source-id: bd20253a93c978e385299ecb8d1b73abf3c5e71b