[flow][refactor] Avoid using `Modulename.String` to represent hardcoded resolved module in insert_type_utils
Summary:
In a future world where Flow allows multiple haste namespaces to co-exist, we need to tag each haste module with the namespace, which should be represented by `Modulename.String _`. However, currently we are abusing this here. This diff replaces it with ad-hoc variant instead. Ideally, the code here needs to be rewritten so that these special cases are not needed.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D64492937

fbshipit-source-id: 77de5525a66490bf982dcccc29a546607c615482