[flow][tests] Fix serious type sig resolution bug wrt to component type
Summary:
It turns out that due to this scoping bug in type sig, generic component types are never correct...

Before the fix, `component(...Props)` for a upper-scope generic `Props` will incorrectly resolve it to non-generic definitions of `Props`. Most of the times, it resolves as global type variable, which silently becomes any in type sig... It shows up in the added test in `component_type` test suite as a weird `underconstrained-implicit-instantiation` error.

The fix is very simple: correctly pass along the `xs` set of generic tparams.

Changelog: [fix] Fix a bug where scope analysis of exported component type is done incorrectly, which causes generics mentioned within component type to be incorrectly typed as any.

Reviewed By: panagosg7

Differential Revision: D64523653

fbshipit-source-id: 1c5571f062f9d21b9f567ef0c4b63011544eb3b6