[flow][multi-haste] Make haste map in shared memory keyed by module name + hardcoded haste namespace bitset
Summary:
In this diff, I implemented foundational support for flow to understand the co-existence of multiple haste namespaces, so that two large codebases with separate haste namespaces can be unified under a single flow root.

To model this, we will change the current model of a haste module being uniquely identified by the globally unique module name, to be uniquely identified by the module name + a namespace set (which can be efficiently encoded in a bitset). For example:

```
- web
  - A.js -> key=A:0b01
- shared
  - B.js -> key=B:0b11
- mobile
  - A.js -> key=A:0b10

(web + shared and shared + mobile
are two previously separate flow roots)
```

Under this scheme, we will be able to tolerate the same module name under separate haste namespaces. Module resolution and conflict detection will become more complex, but I will explain it in future diffs. This diff focuses solely on getting the namespace bitset into shared memory.

This diff introduces `Haste_module_info` that contains both name and namespaces. `HasteMap` is now keyed by this structure instead of just string. In the shared memory, we will store the bitset as a 64-bit tagged int, right next to the module name string. When we previously read the module name out of the shared memory, we will now read both the module name and the bitset.

Not every future issue has been resolved in this diff. Notably, all of the current `Haste_module_info.of_module_name` is problematic, but I will resolve these in future diffs, since that's actually where the core new complexity of module resolution that's a result of this new model.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D64450580

fbshipit-source-id: ac79fcf6c4f02cbd57f702136e1bd928fdee27ee