[flow][type-sig] add support for Object.freeze({[props]} as const)
Summary:
Treat `Object.freeze({[props]} as const)` as `{[props]} as const` since the latter provides stronger guarantees that subsume those of Object.freeze().

This is already the behavior in statement.ml. This diff adds support in type-sig, where this pattern was a signature verification error before this change.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D64917509

fbshipit-source-id: 3c3520d20328660f860e14553fa39b51c763b5aa