[flow] allow creation of objects with computed properties of string union type
Summary:
Before this change, creating objects with computed properties of union string types was not allowed ([invalid-computed-prop] error). This was done in D18089625 mostly for performance reasons.

This diff applies a more permissive approach where we allow unions of strings as computed keys, and just treat them as the general `string` type.

Changelog: [fix] Flow allows unions of strings to be used as the key in creation of objects with computed properties.

Reviewed By: SamChou19815

Differential Revision: D65118717

fbshipit-source-id: 8f0332bb6b388be462f749d3dda235f00719a1cb