[flow][glean] optimize "member declaration reference" computation
Summary:
Passes an "allowed_prop_names" parameter to the normalizer's `ExpandMembersConverter`. This list of names are going to be the only ones considered when the normalizer expands object members.

The goal here is to avoid doing unnecessary normalization work in clients that just access a specific property on the output of `Ty_members.extract`. (See `GleanRunner.extract_member_def`.)

I tried another alternative, which is to call `GetPropT` directly on the input instead of going through the "expand members" procedure, but this introduced some discrepancies in the returned locations. This approach in this diff does not introduce a diff in the output.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D65256740

fbshipit-source-id: bf518aa5ee2912d516068706fcba65ad81e510b9