Fix fails_early_exn test in OSS CI (#9221)
Summary:
It appears to me that somehow the exception is immediately thrown instead waiting until `let%lwt _ = LwtUtils.both_result p1 p2 in` to throw, so I rewrite it slightly to make `Lwt.wakeup_exn r2 (Failure "did not work");` be within the try block.

Not sure whether it still tests the original intent, but that's the runtime behavior now.

Pull Request resolved: https://github.com/facebook/flow/pull/9221

Reviewed By: jbrown215

Differential Revision: D65365972

Pulled By: SamChou19815

fbshipit-source-id: d0b0532d39d78238ad5611fce6d0944722b9faa3