[flow] Implement partial support for react 19's ref-as-prop model
Summary:
In this diff, I implemented the partial support for react 19's ref-as-prop model. The desired behavior under this flag is:

- for jsx checking, we will no longer filter away the ref prop when flowing into expected props, and check the ref prop against `React.RefSetter<void>`. Instead, we will just do `jsx_props ~> fn_component_props`.
- for subtyping, the ref prop in fn component props will be checked against the instance information in component types.

To implement both ideas above, we extend the current unconditional ref filtering in ObjKit with a configurable option of `FilterRef | KeepRef | AddRef of Type.t`. In the comments of react_kit, I explained why the `ref_manipulation` strategy is chosen. In addition, for `FunT ~> AbstractComponentT` subtyping, we will do a `React.ConfigCheck` with both props and instance information from `AbstractComponentT`.

There are some tricky cases with implicit instantiations, so I documented this in the comment why certain strategies are chosen. In cases where we cannot make a reasonable decision because things are unresolved, I pick a strategy that's close to the current inference behavior to minimize error diff.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D64992339

fbshipit-source-id: 32863db7c867bf3e3e917c9e66257b3061de59a6