[flow][match] Basic type checking for body of match expresssion
Summary:
Changelog: [internal]

Some basic support for type checking the match expression as an expression.
The resulting type is the union of the types of the bodies of each clause, abnormal returns excluded.
If every clause body throws, or if any `if` guard throws, the entire expression is considered to throw.

Reviewed By: SamChou19815

Differential Revision: D64840322

fbshipit-source-id: 022baf1099e81500a2a6084ca15c53a81279f556