[flow] Stop searching for unresolved upper bounds for autocomplete
Summary:
We have a few type aware autocomplete for literals. In the pre-LTI world, we will generate unconstrained tvar at these locations with the hope of getting some upper bounds through later interactions, and then we use the upper bound for the type-aware autocomplete.

In the LTI world, this scheme no longer works, and we now explicitly use type hint mechanism to generate an expected type at that location. This expected type is already resolved, so the cleanup here is long overdue.

This diff stops searching for unresolved tvar and explicitly do nothing if we somehow encounter them. I already renamed the upper_bound with expected_type to clear some confusion. I noticed this when I'm trying to change `ReposLowerT` to use named arguments.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D65554633

fbshipit-source-id: ac1c804df4986b86f91d1ee10686809e4198905a