[flow] Clear eval_repos_cache at the same point we clear constraint cache
Summary:
This cache has similar characteristics with constraint cache: LTI breaks cycles between components, so the caches are not useful to break cycles across component boundaries. Therefore, we can also reset this cache.

I originally hope that it can help perf by reducing the number of the items in the cache that would otherwise never decrease in size. It turns out that it's perf neutral, but it's still good to know that it's unnecessary to make the cache have the lifetime of the entire file's type checking.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D65557799

fbshipit-source-id: b8ca1bb26e605c84d5d9c9875050a278c3ec26b6