[flow] Stop caching call related use_ts in constraint cache
Summary:
There is little point caching them since it shouldn't be repeated and doing comparisons on all the arguments are just wasteful.

However, this seems to be a perf neutral change, but it's still good to check in so that we are sure that these caches are not necessary.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D65612795

fbshipit-source-id: 83ef973068d7a908f5ba411ea3140bfc7ffcb8f6