[flow][match] Add better parse error for multiple match args
Summary:
We don't allow multple match args (parsing as AssignmentExpression rather than Expression), to reserve the future possibilty for allowing multiple args (where they would mean the same as `match ([a, b]) {}`.

This is currently a parse error, but working on the Hermes parser version, I realized we can have a better parse error here. Rather than attempting in parse as a `match` call and then failing on the `{`, we can have a better parse error when a single argument is not supplied.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D65851645

fbshipit-source-id: 5025e0db5c14cd9b46b7c7fc91e5a46b1c911a74