[docs] Fix @flowtyped React sample (#9227)
Summary:
Hello,

if we follow the `flowtyped` library definitions example for `react` at https://flow.org/en/docs/libdefs/creation/#toc-declaring-a-module-in-at-flowtyped it leads to an `Cannot resolve name ReactSetStateFunction .Flow(cannot-resolve-name)` error.
This is due to the namespace import. So the correct use of the type would be `React.ReactSetStateFunction`.  Which feels a bit redundant.
I assume those library definitions are not global, so took the liberty to remove the `React` prefix in the examples.
Please tell me if it fits the initial intention, otherwise I can perform changes.

Pull Request resolved: https://github.com/facebook/flow/pull/9227

Reviewed By: SamChou19815

Differential Revision: D65993713

fbshipit-source-id: f3e0bc7effc37c099e586e63183e074f15cc06c1