[flow] Do not add and use synthetic multiplatform synthetic imports for libdef files
Summary:
I see crash like the following in some of the logs:

```
 Failure("remove_old_dependent failed: /libdef/dir/foo.js.flow is not in dependents of module /libdef/dir/foo.web")
```

It's hard to consistently repro, but I believe it's due to the edge case of multiplatform support. Since this conformance check doesn't make sense for libdefs anyways, we should just not add and use these synthetic imports.

Changelog: [fix] Fixed a potential crash when experimental multiplatform support is enabled.

Reviewed By: gkz

Differential Revision: D66238215

fbshipit-source-id: 2e00bcc232c84b9fa5c6a280b65623b994484984