[RFC][flow] Stop accidental exact react element through reinterpretation of AbstractComponent in type ref
Summary:
## Motivation

In this diff, I aim to close the last loophole of creating exact react element without feeling bad: through the shorthand syntax.

From what I observed, people might use this to specify exact react element type because:

1. less character to type `Foo` than `renders Foo`, and more specific type always feels better than more general ones
2. clearly accidental: thinking that `renders Foo | Bar` means `renders (Foo | Bar)`
3. intentional: I really want to use exact react element type so I can inspect jsx or cloneElement, but don't want to make the code look bad.

The usage will only increase if we don't lock down fast.

## Implementation

I reinterpret the shorthand syntax to be the new `React$RendersExactly` opaque type, upper bounded by `React.Node` to maintain the nice property that all arguments to render type are subtypes of `React.Node`.

This type cannot be inhabited by any values created in the userland, so it's practically unusable unless it's used as arguments of render types, so it will finally prevent the easy creation of exact react element types.

I updated the render type normalization to additionally recognize this new opaque type, and our builtin utility types designed to work with shorthand syntax to include this type.

## Alternatives

Why don't we create a type-aware lint to prevent this? I think the lint is less ideal because it will restrict what users can do with the shorthand syntax. Any lint that tries to prevent this would fall into the shape of only allowing the syntax in specific context. It would prevent people to create utility types similar to our builtin PropsOf. It will also prevent people from exporting a union of such shortand syntax to be used with render types.

Reviewed By: panagosg7

Differential Revision: D60634061

fbshipit-source-id: a587ae54a011375aa8639a6d7866394f1e35ceab