[flow][refactor] Remove before adding master_cx
Summary:
The shared memory heap will not update if the key already exists. Previously, it's not needed because we never attempt to write to this entry again in the same server. However, when we switch to the model of reinit instead of restarting the server when libdef changes, it will start to matter.

This diff does this first unconditionally.

Reviewed By: panagosg7

Differential Revision: D66465439

fbshipit-source-id: 2bb082d6caa9c5111a642a2cdcf9a4c4d55fade3