[flow] Recover from libdef changes in libdef files with a reinit with a scheduled full check
Summary:
Currently, if you edit a libdef file, it will cause the server to exit. When the server gets restarted, it will enter lazy mode with an always unusable saved state, so it will enter the unsound fallback lazy mode, where we assume that the lazy server always starts in a clean state, so it will just parse all files and check no files.

Two issues together cause the unsoundness: exiting the server and the unsound lazy mode. In this diff, I will only partially fix the problem, which i believe is the most common one: we will not exit the server when a libdef file changes, we will recover.

When libdef has changed, we will schedule a special kind of recheck (`RequiredFullCheckReinit`, as opposed to `NormalUpdates`). When rechecker encounter this, it will trigger a re-init, where we will run an init similar to the one with saved-state init, and then schedule a recheck that include all the files.

The change is gated with the flag introduced in the previous diff, in case this breaks things in unexpected ways.

Changelog: [internal] New behavior is off-by-default for now.

Reviewed By: panagosg7

Differential Revision: D66314602

fbshipit-source-id: df9480588ecdaeff0246f8f61c7cfd00973c4e90